contact-backend/requirements.txt
flask
python-dotenv
flask_cors
